requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.10.5
google-generativeai>=0.3.0
python-dotenv>=1.0.0
websockets==12.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
alembic>=1.13.0

# Optional: Authentication (uncomment if using auth.py)
# python-jose[cryptography]>=3.3.0
# passlib[bcrypt]>=1.7.4
